With general::hide_opers off, behaviour reverts back to generic charybdis - users may see opers. This is the behaviour most IRC networks will want.
This commit will be followed in future, possibly with the 0.2.4 release, with
a split. hide_opers_from_opers will be total operhide, which is ircd-seven
hide_opers behaviour. Current ircd-seven behaviour is to hide opers from non-
opers, and if hide_opers is on, also from opers. hide_opers in 0.2.4 will hide
opers from non-opers.